# apps/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Char AI API",
    "description": "AI services API for speech-to-text transcription, LLM chat completions, and subscription management",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/subscription/delete-account": {
      "delete": {
        "tags": [
          "subscription"
        ],
        "operationId": "delete_account",
        "responses": {
          "200": {
            "description": "Account deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/support/chatwoot/conversations/{conversation_id}/messages": {
      "get": {
        "tags": [
          "chatwoot"
        ],
        "operationId": "get_messages",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "description": "Conversation ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "source_id",
            "in": "query",
            "description": "Contact source ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Chatwoot API error"
          }
        }
      },
      "post": {
        "tags": [
          "chatwoot"
        ],
        "operationId": "send_message",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "description": "Conversation ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Chatwoot API error"
          }
        }
      }
    },
    "/sync/token": {
      "post": {
        "tags": [
          "sync"
        ],
        "operationId": "create_credentials",
        "responses": {
          "200": {
            "description": "Short-lived CloudSync credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudsyncCredentials"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Meetspace Pro subscription required"
          },
          "502": {
            "description": "Credential issuer unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloudsyncCredentials": {
        "type": "object",
        "required": [
          "databaseId",
          "token",
          "expiresAt",
          "workspaceId"
        ],
        "properties": {
          "databaseId": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          }
        }
      },
...
      "DeleteAccountResponse": {
        "type": "object",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "messageType": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "messageType": {
            "type": "string"
          },
          "sourceId": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      }
    }
  },
...
}